fix: fix database migration in production

+++ src/lib/db/seed.ts
@@ -2,6 +2,10 @@ import { db } from "."
 import { quotes, records } from "./schema"
 
 export async function seedDatabase() {
+  if (process.env.NODE_ENV !== "development") {
+    return
+  }
+  
   console.info("DELETING OLD SEED DATABASE")
   await db.delete(quotes)
   await db.delete(records)
